+++ Configuration/TCA/tx_realty_documents.php
@@ -1,88 +1,16 @@
 <?php
 defined('TYPO3_MODE') or die('Access denied.');
 
-$tca = [
+return [
     'ctrl' => [
-        'title' => 'LLL:EXT:realty/Resources/Private/Language/locallang_db.xlf:tx_realty_documents',
-        'label' => 'title',
-        'tstamp' => 'tstamp',
-        'crdate' => 'crdate',
-        'cruser_id' => 'cruser_id',
-        'languageField' => 'sys_language_uid',
-        'transOrigPointerField' => 'l18n_parent',
-        'transOrigDiffSourceField' => 'l18n_diffsource',
-        'default_sortby' => 'ORDER BY sorting',
         'delete' => 'deleted',
         'hideTable' => true,
-        'enablecolumns' => [],
-        'iconfile' => 'EXT:realty/Resources/Public/Icons/Document.gif',
     ],
     'columns' => [
-        'sys_language_uid' => [
-            'exclude' => 1,
-            'label' => 'LLL:EXT:lang/locallang_general.xlf:LGL.language',
-            'config' => [
-                'type' => 'select',
-                'renderType' => 'selectSingle',
-                'foreign_table' => 'sys_language',
-                'foreign_table_where' => 'ORDER BY sys_language.title',
-                'items' => [
-                    ['LLL:EXT:lang/locallang_general.xlf:LGL.allLanguages', -1],
-                    ['LLL:EXT:lang/locallang_general.xlf:LGL.default_value', 0],
-                ],
-                'default' => 0,
-            ],
-        ],
-        'l18n_parent' => [
-            'displayCond' => 'FIELD:sys_language_uid:>:0',
-            'exclude' => 1,
-            'label' => 'LLL:EXT:lang/locallang_general.xlf:LGL.l18n_parent',
-            'config' => [
-                'type' => 'select',
-                'renderType' => 'selectSingle',
-                'items' => [['', 0]],
-                'foreign_table' => 'tx_realty_documents',
-                'foreign_table_where' => 'AND tx_realty_documents.pid=###CURRENT_PID### AND tx_realty_documents.sys_language_uid IN (-1, 0)',
-                'default' => 0,
-            ],
-        ],
-        'l18n_diffsource' => [
-            'config' => [
-                'type' => 'passthrough',
-            ],
-        ],
         'object' => [
             'config' => [
                 'type' => 'passthrough',
             ],
         ],
-        'title' => [
-            'exclude' => 0,
-            'label' => 'LLL:EXT:realty/Resources/Private/Language/locallang_db.xlf:tx_realty_documents.title',
-            'config' => [
-                'type' => 'input',
-                'size' => 30,
-                'eval' => 'required',
-            ],
-        ],
-        'filename' => [
-            'exclude' => 0,
-            'label' => 'LLL:EXT:realty/Resources/Private/Language/locallang_db.xlf:tx_realty_documents.filename',
-            'config' => [
-                'type' => 'group',
-                'internal_type' => 'file',
-                'allowed' => 'pdf',
-                'max_size' => 4000,
-                'uploadfolder' => 'uploads/tx_realty',
-                'size' => 1,
-                'minitems' => 0,
-                'maxitems' => 1,
-            ],
-        ],
-    ],
-    'types' => [
-        '0' => ['showitem' => 'sys_language_uid, l18n_parent, l18n_diffsource, title, filename'],
     ],
 ];
-
-return $tca;